fixed loads, fixed load viewer, and improved load viewer.
Load box had issues with load_array being iteratted over by a for loop, i fixed that. the load array was being set to [0] before calculating time, I fixed that. and now the time changes to 0.000 if you delete the load and if the time is 0 it doesn't show it inside of the load viewer. :)